notifier/pyproject.toml
[project]
name = "notifier"
version = "0.1.0"
description = "This notifies interested parties of events relating to Zetachain."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dateparser>=1.2.0",
    "feedparser>=6.0.11",
    "flask>=3.1.0",
    "humanize>=4.11.0",
    "jsonpickle>=4.0.0",
    "loguru>=0.7.2",
    "python-dateutil>=2.9.0.post0",
    "python-telegram-bot>=21.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "telegram>=0.0.1",
]
